Xử lý mua bán, đặt tháp.
- Xóa bớt comment.
- Hoàn thành reset game khi kết thúc.
- Tối ưu một số đoạn code.
- Viết thêm hàm vào MediaManager.
- Thêm một ít hằng số.
- Thêm một số asset âm thanh.

+++ src/main/java/game/tower/Tower.java
@@ -9,7 +9,7 @@ import javafx.scene.image.ImageView;
 import javafx.scene.paint.Color;
 
 public abstract class Tower extends AbstractTile {
-    protected int fireRate = 0;
+    protected double fireRate = 0;
     protected int fireRange = 0;
     protected Bullet bullet = null;
     protected Image gunImg = null; // super.image is base image.
@@ -101,11 +101,11 @@ public abstract class Tower extends AbstractTile {
         return bullet;
     }
 
-    public int getFireRate() {
+    public double getFireRate() {
         return fireRate;
     }
 
-    public void setFireRate(int fireRate) {
+    public void setFireRate(double fireRate) {
         this.fireRate = fireRate;
     }
